requirements.txt
click==7.1.2
confluent-kafka==0.11.4
kafka-python==2.0.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==1.4.1
kazoo==2.4.0
MarkupSafe==1.1.1
Werkzeug==1.0.1
WTForms==2.3.3
email-validator==1.1.2
Flask-WTF==0.14.3
numpy==1.14.1
opencv-python=3.4.0.12
six==1.11.0
tabulate==0.8.2
wheel==0.26.0

export FLASK_ENV="DEVELOPMENT"